+++ client/src/pages/api/support/file/delete.ts
